% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_gec.R
\name{annotate_gec}
\alias{annotate_gec}
\title{annotate for grammar error correction}
\usage{
annotate_gec(input, force_handwriting = FALSE)
}
\arguments{
\item{input}{path to file or character vector of text}

\item{force_handwriting}{whether to skip the attempt to parse the text locally}
}
\value{
a list containing an imagemagick image and corrected text data.frame.
 If only text is provided, the imagemagick image will simply return NULL
}
\description{
wraps annotate_text, annotate_pdf_digital and annotate_pdf_handwriting
}
