% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pdf_digital.R
\name{annotate_pdf_digital}
\alias{annotate_pdf_digital}
\title{poppler pdf text pipeline}
\usage{
annotate_pdf_digital(pdf_file)
}
\arguments{
\item{pdf_file}{file for correction}
}
\value{
a list: $images = imagemagick image pointer, corrections = table output from `annotate_text`
}
\description{
takes a file, returns a table of corrected sentences and annotated images
}
