% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pdf_handwriting.R
\name{annotate_pdf_handwriting}
\alias{annotate_pdf_handwriting}
\title{handwriting api pipeline}
\usage{
annotate_pdf_handwriting(file)
}
\arguments{
\item{file}{file for correction}
}
\description{
takes a file, returns a table of corrected sentences and annotated images
}
