% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_text.R
\name{annotate_text}
\alias{annotate_text}
\title{annotate digital text}
\usage{
annotate_text(digital_text)
}
\arguments{
\item{digital_text}{character vector}
}
\value{
a tibble
}
\description{
takes a character vector and runs grammar error correction on it
requires `gec_env` to be used, spacy to be instantiated, and sentence.py to have been sourced
}
\examples{
#annotate_text("Ma, does ye have any potatos?")
}
