% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_conda_gec_env.R
\name{create_conda_gec_env}
\alias{create_conda_gec_env}
\title{create grammar error correction python environment}
\usage{
create_conda_gec_env(
  packages = NULL,
  forge = TRUE,
  channel = character(),
  conda = "auto",
  python_version = 3.8
)
}
\arguments{
\item{packages}{packages to include beyond those specified in requirements.txt (none should be necessary)}

\item{forge}{whether to use forge}

\item{channel}{specific channels to prefer}

\item{conda}{which conda to use}

\item{python_version}{which python version to use - defaults to 3.8 for guaranteed compatibility}
}
\value{
a python environment with requisite packages names "gec_env"
}
\description{
create grammar error correction python environment
}
