% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_digital_text.R
\name{extract_digital_text}
\alias{extract_digital_text}
\title{extract digital text}
\usage{
extract_digital_text(data)
}
\arguments{
\item{data}{either a dataframe with column "token" or a character vector to be squished down}
}
\value{
a character vector of length one with all the text
}
\description{
simply turn the tokenized text into regular text.
}
\examples{
extract_digital_text(data = c("A bunch of text", "that hasn't been connected yet." ))
extract_digital_text(data.frame(token = c("A", "bunch", "of", "'", "text", "'", ".")))
}
