% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcv_auth.R
\name{gcv_auth}
\alias{gcv_auth}
\title{authentication wrapper}
\usage{
gcv_auth(client_id, client_secret, service_token)
}
\arguments{
\item{client_id}{your specific client id}

\item{client_secret}{your specific client secret}

\item{service_token}{your specific service token}
}
\value{
an authenticated session
}
\description{
authentication wrapper
}
