% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcv_check_filetypes.R
\name{gcv_check_filetypes}
\alias{gcv_check_filetypes}
\title{convert files for gcv if necessary}
\usage{
gcv_check_filetypes(file)
}
\arguments{
\item{file}{file to be converted}
}
\value{
tempfiles and their paths
}
\description{
if a pdf is provided, it will be converted into 300 dpi pngs
in a temp directory to be sent to the API
}
\details{
if an image file is provided, it will return that

otherwise, it will return NULL with a warning
}
