% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcv_reorder_text.R
\name{gcv_reorder_text}
\alias{gcv_reorder_text}
\title{get returned text in (probably) the correct order}
\usage{
gcv_reorder_text(data)
}
\arguments{
\item{data}{the result of `gcv_xy`}
}
\value{
a reordered tibble with additional information
}
\description{
- removes likely duplicates
- orders the text in logical lines, allowing for some amount of skew
- standardizes the tokenization as far as possible
}
\details{
the order_id this produces is used in downstream tasks
}
