% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcv_xy.R
\name{gcv_xy}
\alias{gcv_xy}
\title{extract relevant x and y location annotations from google cloud vision}
\usage{
gcv_xy(data)
}
\arguments{
\item{data}{result of `googleCloudVisionR::gcv_get_image_annotations`}
}
\value{
table with additional columns
}
\description{
the API provides 4 x and y coordinates each, we just want to render that as a simple rectangle
}
