% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gec_env.R
\name{gec_env_create}
\alias{gec_env_create}
\title{create grammar error correction python environment}
\usage{
gec_env_create(
  envname = "gec_env",
  packages = NULL,
  forge = TRUE,
  channel = character(),
  conda = "auto",
  python_version = 3.9
)
}
\arguments{
\item{envname}{name of environment if you don't want to use the default}

\item{packages}{packages to include beyond those specified in requirements.txt (none should be necessary)}

\item{forge}{whether to use forge}

\item{channel}{specific channels to prefer}

\item{conda}{which conda to use}

\item{python_version}{which python versionexists to use - defaults to 3.8 for guaranteed compatibility}
}
\value{
a python environment with requisite packages names "gec_env"
}
\description{
create grammar error correction python environment
}
