% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gec_env.R
\name{gec_env_use}
\alias{gec_env_use}
\title{use gec env and load python grammar error correction function}
\usage{
gec_env_use(envname = "gec_env")
}
\arguments{
\item{envname}{name of environment if you don't want to use the default}
}
\value{
spacy initialized python environment and access to `correct_sentences.py`
}
\description{
use gec env and load python grammar error correction function
}
