% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rotation.R
\name{orient_files}
\alias{orient_files}
\title{Conform images to exif metadata}
\usage{
orient_files(files)
}
\arguments{
\item{files}{paths to files}
}
\value{
re-oriented files
}
\description{
[Exif metadata](https://en.wikipedia.org/wiki/Exif) records the orientation of images. Before passing
images to the handwriting API, you will want to make sure that the x and y axes are the way the picture-taker intended.
}
\details{
This is going to be fairly slow, as it is reading and writing images to/from (temp)files, so at some point there should be
some reading of metadata to find out whether it needs to be done or not.
}
