% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppler_reorder_text.R
\name{poppler_reorder_text}
\alias{poppler_reorder_text}
\title{Extract tokenized dataframe from pdf}
\usage{
poppler_reorder_text(pdf_file)
}
\arguments{
\item{pdf_file}{the file to be read}
}
\value{
data.frame of tokens and locations
}
\description{
takes a readable digitally produced pdf file and returns the equivalent reuslt of `gcv_reorder_text`
}
