% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_annotations.R
\name{prep_annotations}
\alias{prep_annotations}
\title{Prepare annotations for imagemagick}
\usage{
prep_annotations(annotated_text, digital_text_df)
}
\arguments{
\item{annotated_text}{output of `annotate_text`}

\item{digital_text_df}{output of `gcv_reorder_text` or `poppler_reorder_text`}
}
\value{
data.frame with annotation data
}
\description{
Take the corrected text and the tokenized df with locations, and find where to put
what kind of annotation
Gives a warning when row counts do not quite match up (off-by-one or more errors are present)
}
