% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_conda_gec_env.R
\name{use_gec_env}
\alias{use_gec_env}
\title{use gec env and load python grammar error correction function}
\usage{
use_gec_env(envname = "gec_env")
}
\arguments{
\item{envname}{name of environment if you don't want to use the default}
}
\value{
spacy initialized python environment and access to `correct_sentences.py`
}
\description{
use gec env and load python grammar error correction function
}
