% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdiff.R
\name{wdiff}
\alias{wdiff}
\title{wdiff behaviour in R}
\usage{
wdiff(original, correction, sep = " ")
}
\arguments{
\item{original}{character vector of original sentences}

\item{correction}{character vector of corrected sentences}

\item{sep}{separator for tokenization (use " " to replicate wdiff, or "" for character by character diff)}
}
\value{
a character vector summarising diffs between the two character vectors
}
\description{
replicate the simple wdiff output without writing tempfiles our calling the command line
}
\examples{
wdiff("I haben't idea about that", "I have no idea about that.", "")
}
