% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdiff.R
\name{wdiff_}
\alias{wdiff_}
\title{wdiff behaviour in R}
\usage{
wdiff_(original, correction, sep = " ")
}
\arguments{
\item{original}{original sentence}

\item{correction}{corrected sentence}

\item{sep}{separator for tokenization (use " " to replicate wdiff, or "" for character by character diff)}
}
\value{
a character vector length 1 summarising diff
}
\description{
replicate the simple wdiff output without writing tempfiles our calling the command line
}
